requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.7.0
anthropic==0.28.0
openai==1.35.0
chromadb==0.5.0
networkx==3.3
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
